pyproject.toml
[project]
name = "django-restapi-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.3",
    "django>=5.2.1",
    "django-celery-results>=2.6.0",
    "djangorestframework>=3.16.0",
    "drf-yasg>=1.21.15",
    "redis>=7.4.0",
    "requests>=2.32.4",
    "weasyprint>=68.1",
]
